pyproject.toml
[project]
name = "alphabacktest"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "matplotlib>=3.9.4",
    "openbb>=4.1.3",
    "pandas>=2.2.3",
    "pyyaml>=6.0.2",
    "tqdm>=4.67.1",
]
